feat: retire standalone eligibility-check; on-page apply site-wide
Roll the on-page apply pattern into every charity funnel and retire the
separate /eligibility-check/ wizard page (the hand-off to it was jarring).

- Extract the inline "Help me choose" guide into a reusable
  ui/HelpMeChoose component (native <details> + native radios, radio
  group scoped with useId so multiple instances never collide). ApplyOptions
  now composes it.
- Hosting: ReadyToGetStarted gains an #apply anchor and HelpMeChoose under
  the two apply buttons; Hero and HowToApply CTAs scroll to #apply instead
  of linking the wizard.
- Domains: Dear-Prospective gains the #apply anchor + HelpMeChoose; Hero
  and Order step 1 scroll to #apply; the stale eligibility hint is dropped
  from Get-New-Website (its apply buttons stay).
- Homepage Our-Programs and the badge / charities-we-support inbound links
  now point at /help-for-charities/ (the apply hub).
- Retire /eligibility-check/: delete the route, 301 it to
  /help-for-charities/ in .htaccess, and drop it from the sitemap, guides
  list, search index (regenerated), routes list, and the wizard e2e spec.
  DecisionWizard stays (still used by the volunteer quiz).
