+++ src/components/domains/Order-Your-Domain/index.tsx
@@ -7,11 +7,11 @@ const HowToOrderDomain = () => {
     {
       number: 1,
       title: 'Step 1',
-      description: 'Check your eligibility, then start onboarding',
-      linkText: 'Check eligibility',
+      description: 'Complete FFC onboarding',
+      linkText: 'How to apply',
       innerbg: 'bg-[#8A6400]',
       outerbg: 'bg-[#fff]',
-      linkUrl: '/eligibility-check/',
+      linkUrl: '#apply',
     },
     {
       number: 2,
