+++ tests/routes.ts
@@ -79,7 +79,6 @@ export const siteRoutes = [
   { route: '/volunteer-roles', name: 'Volunteer Roles' },
   { route: '/volunteer-onboarding-guide', name: 'Volunteer Onboarding Guide' },
   { route: '/contribute', name: 'Contribute on GitHub' },
-  { route: '/eligibility-check', name: 'Eligibility Check' },
   { route: '/volunteer-quiz', name: 'Volunteer Quiz' },
   { route: '/matching-gifts', name: 'Matching Gifts' },
   { route: '/badge', name: 'Powered-by Badge' },
